services/auth-service/requirements.txt
fastapi==0.115.5
alembic==1.14.0
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
prometheus-client==0.21.0
uvicorn==0.30.6
